feat: add inline bit buffer implementation that utilizes the backing ByteBuf instance to store the bit reader and writer indices.
The benchmarks demonstrate a 10.4% performance increase when writing a small amount of bits (30 bytes in the test), which can be attributed to memory locality between the underlying buffer and the indices.
Allocated BitBuf implementation however still outperforms inline bit buffer. InlineBitBuf performs 18.3% worse than the allocated counterpart when writing a large amount of data (40,000 bytes).

In both benchmarks, Netty's checks are disabled as that causes a further hit to the inline implementation - the checks would be disabled in production anyhow.