feat: add support for an unsafe long-backed bit buffer, used to cache high and low res bitcodes for player info.
This implementation is unsafe to use as it does not do any bound checking. The payload is stored in a 'long' primitive type, meaning a maximum of 64 bits of data could ever only be written. This works for our player info needs as the maximum every written per low/high res caps out in the 40s.