feat: add high level logging support for rsprot, with js5 being separated out under its own flag as it is super spammy;
This is helpful for testing that everything functions as expected, it should not, however, ever be used in production.
The performance penalty here is negligible as all the log calls are wrapped in a simple `if (boolean)` check, ensuring nothing gets executed if the boolean is false, nor will the message be constructed.