feat: introduce server prot categories
In OldSchool RuneScape, the packets get categorized in two:
One category which includes most of the engine packets, but also some content-created packets.
One category which includes most of hte content-level triggered packets.

Below is a list of all the known __high__ priority engine packets, the ones sent in the first half. Note that the list here corresponds to the order in which these packets are sent on login, so the relative order between what is defined here is also correct:
- RebuildLogin
- RebuildNormal
- RebuildRegion
- UpdateSiteSettings
- ChatFilterSettings
- MessageGame
- VarpReset
- VarpSmall
- VarpLarge
- VarClanDisable
- PlayerInfo
- NpcInfoSmall
- NpcInfoLarge
- UpdateZoneFullFollows
- UpdateZonePartialFollows
- UpdateZonePartialEnclosed
- ZoneProt (all implementations)
- UpdateInvFull
- UpdateInvPartial

Below is a list of all the known __low__ priority content packets, the ones sent in the second half. Note that the list here corresponds to the order in which these packets are sent on login, so the relative order between what is defined here is also correct:
- RunClientScript
- IfSetText
- IfOpenTop
- IfOpenSub
- IfSetEvents
- IfInitialState
- CamReset
- SetPlayerOp
- UpdateStat
- UpdateRunWeight
- UpdateRunEnergy
- ResetAnims
- MinimapToggle
- UpdateFriendChatChannelFullV2
- UpdateTradingPost
- UpdateStockMarketSlot
- UpdateFriendList
- ServerTickEnd

Any packet which is not listed in the above two lists has been guessed based on other similar packets. As an example, it is safe to assume that all IF_* packets belong in the low priority packets group, as two similar-state packets cannot belong in separate groups - that could result in problems, like IF_SETOBJECT being sent before IF_OPENSUB, which would result in the IF_SETOBJECT packet doing nothing. Educated guesses were made for all such packets, along with a few other tests to ensure the correctness of these lists.
Furthermore, upon implementing, a function to pass a priority override will be provided, allowing servers to override any priorities which may have been misconfigured by this library.