feat: support beta world logins
This commit adds support for the special login procedure that happens
if the client attempts to log into a world that has a beta flag,
as only the partial CRCs are sent in this scenario.
The server then has to respond with opcode 2 without any other
data besides that. Afterwards, the client will send the remaining
CRCs in a separate response to the server, which the server can
merge into a single complete block of CRCs, and fully validate
the login.

While the above described structure isn't the intended purpose,
the implementation of beta worlds is poor on Jagex's behalf,
and does not actually prevent anything. As a result of this,
just to make the entire thing function, we automatically request
the remaining CRCs before notifying the server of anything.