fix: correctly deallocate info buffers
This ensures that if either of the infos was in fact built,
but the deallocate function is called before the respective
toPacket function, we release the allocated buffer.
This can happen due to exceptions or just abnormal process flow,
in any case, we make the assumption that if an info object is
deallocated, it will not be written into the channels from that
point onward, and as such, it is safe to release the buffers related
to them.