feat: throttle prefetch responses when not on login screen
This adds a fairly complex mechanism that throttles prefetch requests
when the user is not on login screen (excluding the initial loading
screen!). In doing so, we get significantly faster initial load times,
and anyone that happens to log into the game while the cache isn't
yet fully downloaded will not experience significant loading screens
due to the pipeline being filled by previous prefetch requests.
In some extreme cases, this created up to a 10-second-long loading
screen which could, in theory, mean death for HCIM.
The default configuration is set to download the entire cache in just
under 6 minutes on localhost. In real life, ping will play some
role on-top of that, but it shouldn't be a huge difference, as
the client can make further requests while the server handles the
throttled requests (e.g. server might write 30 responses out, client
can already send in 30 new prefetch requests in that time, long
before server gets through with those initial 200 requests it made).