fix: fix a problem with some extended info blocks not being computed
Fixes an issue where some extended info blocks, such as face pathing
entity were not computed when a new player started observing a NPC.
This was because of an optimization in place to not compute extended
info blocks for NPCs who have no players observing them. This however
backfires as some extended info blocks are kept long-term and
transmitted to any future observers. As a result of this, the extended
info blocks which may be transmitted in the future will now be
computed even for NPCs who have no observers. However, extended info
blocks such as hits, animation and so on, which only apply to the
current cycle will not be computed in such scenarios.