fix: exclude outer 1 tile border of the build area in contains
This is because all clients exclude at least 1 tile border
in build area checks. On java, it is exactly 1 tile border.
On the C++ client, it is actually two tiles, however movement
still renders fine on the second tile in C++, so our code
will only exclude the one tile border.