Add default arg for `supportedClients`
It's probably best to be as easy-as-possible to create an instance for the average consumer. This would have possible perf cost implications if buffers are being calculated for client types that the consumer doesn't end up requiring; however, as of now, this is a non-issue. (given there's only one supported `OldSchoolClientType`)