feat: add opt-in global packet listener
Adds 'global consumers' which are invoked for every single packet coming from the client to the server, after type-specific listeners are invoked.

Potential use cases which prompted this:
- Passing through packets, or some derivative of, to some sort of bot detection system.
- Having a way to mark a player as 'active' when any packet, or any packet from a set of packets, are received.

The latter situation is used in August to clear a 'self disconnection' timer on a player that is added when the server receives a valid login request while the player is already logged in.
- Player is logged in
- Player's client DC's, but server keeps their player in the game
- Player's client sends another login request with correct password
- `Player` in-game marked with self-DC timer for 10 seconds that'll try boot them in most situations
- After 10 seconds if no packets are received for that player (which is the case when the client DC's), the player is booted