feat: add helpful warnings on js5 buffer incompatibility
The first buffer that gets written to the pipeline via JS5 will
now perform a direct buffer compatibility check and warn the user
if the buffers are incompatible (meaning one is direct, the other is
not). In such cases, a relatively expensive copy process has to take
place which can significantly reduce throughput.