fix: npc info crash if extended info filter always returns false
Fixes an obscure bug that leads to the protocol collapsing.
This can be produced by flagging exactly one NPC with extended info,
but rather than writing the real flag + payload, only write a zero.
We use this strategy to avoid our buffer going out of bounds when
the buffer starts to fill up - as the total number of bytes we need
to write out at that point is equal to the number of NPCs' whose
extended info we've yet to write out.
The problem stems from the client's expectations of the data. While
the client iterates the high resolution NPCs, it expects more data
than would be sent if there's a single NPC with extended info which
we skip by writing a zero. The client's expectation is 16+12 bits of
data, however if we do what I described, we only end up sending a
total of 24 bits of data, through the terminator (16 bits, 0xFFFF)
plus the single zero byte indicating the extended info skip.
This corrupts the client's flow, and it never ends up reading the
terminator itself. This in turn leads to the extended info block
reading the extended info flag as 0xFF, which makes NPC info want
to read a lot more data than is intended, as 0xFF is all 8 bits
enabled. This quickly corrupts and crashes.