fix: update timeouts to match the client
The timeouts are as follows:
Initially, when the connection is made, 30 seconds.
If the connection requests JS5, the timeout remains 30 seconds
for the remainder of that connection.
If the connection requests game login, the timeout is increased to
40 seconds.
Once a successful game login takes place, the timeout is decreased
down to just 15 seconds.