Network Traffic Metrics (#33)
* feat: initial network traffic tracking

* feat: add support for a generic network traffic writer

* feat: add support for tracking login blocks

* fix: add bytes suffix to packet snapshots

* feat: add a function to request closing sessions

Additionally, mark down game dc reason as logout when logging out

* feat: add the ability to freeze and unfreeze network traffic tracking

The freezing mechanism only applies to metrics which aren't stateful,
meaning packets, disconnections and so on. It will not halt tracking
connections from various IPs though, as that bit is stateful. If we did
freeze connections, we could run into trouble whereupon unfreezing,
the metrics could go negative, or over-inflate, since it loses context
about what happened during the freezing portion.

* refactor: rename Generic* to Concurrent* as it gives a better indication

Generic name doesn't really say anything, whereas Concurrent gives
some sort of idea of how the implementation is done. In this case,
most of it is Atomic, but there is a tiny amount of synchronization
when the resetting process takes place.

* docs: add documentaiton to channel handlers

* refactor: rename Handler to Monitor for traffic

* refactor: remove total active connections tracking

This is unnecessary since we track per InetAddress anyway. If one
desires, they can sum up all the values in the respective map and
get to the same outcome.

* refactor: rename InetAddressTrafficCounter to InetAddressTrafficMonitor

* refactor: use enum maps for inet address traffic snapshots