feat: allow players to be marked as high pri. relative to one another (v225 only)
Allows players to mark one another as high priority, meaning they will
render that specific player above large crowds of players, even if they
would not ordinarily fit the distance criteria. This is an addition to
the default high priority system in the previous commit, which allows
specific users to be marked as high priority to everyone.

The specific use cases here would be to mark players in combat between
one another as high priority, ensuring both of them can see each-other.
However, to do this properly, the value must also be reset when combat
ends between them. This is just an example use case and Jagex does not
actually do it, but servers may wish. An additional example could be
to always render anyone on your friend list over anyone else, or for
example your own clan members. One needs to be careful with this though,
as both of those systems allow potentially for hundreds of members,
which goes above the typical preferred count of 250.