feat: add an option to opt out of precise npc avatar tracking
Opting out of it shows a performance improvement of about 13%
in a multithreaded environment. The performance impacts drop
significantly when thread contention decreases, in single-threaded
benchmark there's hardly any impact, but still.
This allows servers to opt-out of it if they want to squeeze those
few bits of extra performance out of it and don't need the precise
avatar tracking logic.