fix: url_open packet
The packet broke a few days ago when a change was made to drop invalid
packets safely without corrupting the stream. This however reordered
some code, making it so payloads were written before the opcode.
For all but one packet, this was fine - that one packet is url_open,
which additionally encrypts the payload. Since that update, the payload
was now encrypted before the opcode, throwing the encoding out of order.