feat: add validation to ensure integrity of info packets (225 only)
Adds extra validation to all three info packets, making sure the
previous info packet was successfully written to the session.
This fixes a potential bug where servers might have a flow problem
or invalid tick order for packets, making it possible for certain
packets to just get calculated but never written, which would adjust
the state in RSProt, but not sync it with the client, causing obscure
errors.