Message Size Estimator (#38)
* feat: add support for message size estimating

* chore: disable ktlint expression body enforcing

This setting often collides with explicit API mode and is
more of a preference anyway.

* feat: accurately estimate outgoing game messages

* feat: accurately estimate outgoing login messages

* fix: remove duplicate constant

* fix: override JS5 message size estimate

Additionally, renames limit to length and fixes a bug of treating
the length as limit when XOR key is used (abnormal activity only)

* fix: allocate correctly sized buffers for encoders

* refactor: switch to a custom MessageToByteEncoder implementation

This switch allows us to avoid doing an entire buffer copy whenever
we pass a ByteBufHolder into the encoder. Instead, we then write a
small header packet which contains the opcode & size, and then the
prebuilt byte buf holder, without ever copying it over.

* fix: increment outgoing bytes for byte buf holder packets

* fix: obtain a new slice of byte buf holder contents

Fixes a problem with shared buffers like partial enclosed.