fix: retain base animations assigned in previous calls in same cycle
Ensures that calling npc#setWalkAnims, followed by npc#setRunAnims
will not end up clearing the walk anims, and includes all 4 of the
animations that were intended to be assigned.

In the old code it would wipe any previous calls, as well as only
assign the last animation within a single function call,
as it only took the last animation flag, rather than all the assigned
flags.