refactor: separate hit source and hit recipient for players
This allows the server to pass in differing hitmark types for
the player that dealt the hit, and the one who received the hit.

This is required as max hit hitmarks are only rendered to players
if the damage value is equal to or higher than the player's input
value in game settings.

Example scenario:
Player A max hits on player B for 15 damage.

1. If player A threshold is >= 15 and player B threshold is >= 15,
both players will see the max hit HitMark.
2. If player A threshold is >= 20 and player B threshold is >= 15,
only player B will see the hitmark as max hit. Player A will see it
as the normal lit-up variant instead.
3. If player A threshold is >= 15 and player B threshold is >= 20,
only player A will see the hitmark as max hit. Player B will see it
as the normal lit-up variant instead.
4. If player A threshold is >= 20 and player B threshold is >= 20,
both players will only see the normal lit-up variant.

Note that in these examples, every other player (besides A & B)
will see the hit as the tinted, dark variant.

This API change has only been done for players, as in the case of NPCs,
there's only the player that dealt the damage, so the old implementation
works as is. The implementation is only done for hitmarks, and not
headbars, as headbars have no equivalent, but they do support
the "specific" implementation (only render for a specific player).

The API change is also backwards-compatible. Existing implementations
will simply pass the existing `selfType` value to the `sourceType`,
accomplishing the same old results.