refactor(229): initiating a shutdown now returns a completable future
This provides more control to the server, and safely shuts the service
down, keeping mind of variables that may not have initialized yet,
and ignoring those. Servers can now configure the timeouts as well.
A change that was made, however, is that start() no longer tries to
handle errors and shut down on its own. The implementation was flawed.
It is up to the caller to handle exceptions from start() and invoke
shutdown if they deem fit.