fix: packet groups now auto-partition internally to avoid errors
The partitioning is done as servers cannot really be aware of how
many bytes a packet consumes very well, so it's hard to safely
partition packets server-sided. This change will now use the worst
case estimates for packets and group them in up to 32767 byte lists,
sending each one as its own packet group. If there's an error in the
estimates, the size transmitted is limited to 32767, but the packet
itself can be up to 40,000 bytes in size, giving us a comfortable
7 kilobytes of headroom.