refactor: player hitmark/headbar functions always error-check
This is because these functions utilize something called a smart1or2,
which means the values cannot just under or overflow as they normally
would. Instead these would corrupt and crash.
Therefore, to avoid internal erroring which is harder to debug,
we've made the error-checking for out-of-range values occur at the
call site, erroring any server code that tries to set up a hitmark
or a headbar with invalid values.