fix: close channel if not closed after netty handler exception caught
Ensures that channels get closed after an exception occurs, if it was
not already caught. Previously it relied on server closing it in the
exception handler, but this puts too much responsibility on the user,
given that none of these errors can ever be recovered from. By not
closing, servers become vulnerable to attacks too.