fix: avoid allocating strings for passwords
Avoids allocating strings for passwords, as those tend to
get internalized and are easy to observe in heap dumps. Instead,
char buffers/arrays are used throughout which are overwritten and
erased after respective operations. This gives servers full control
over when the password itself should be erased from memory, and doesn't
require them to wait for garbage collector to do its job.