perf: reduce initial js5 client allocations
Reduces the initial allocations per JS5 client initialized from three
size-200 int arrays down to size-0. This avoids an OOM error that could
take place is a million (or any absurd number like it) connections are
opened against the server.
Furthermore, limits the maximum possible resized array to 201. This is
to avoid it resizing to 255 when it can never get more than 200 (+1)
entries in the array.