feat: enabled so_reuseaddr & set so_backlog to 4096 by default
These settings can be configured by the user in revision 225+.

so_reuseaddr allows for another socket to immediately connect after
the previous one has closed, rather than needing to wait a brief period
before it can connect.

so_backlog is the maximum number of kernel-queued connections that
can be queued up before the kernel starts rejecting them.
While we do increase the default on our end to 4096, it is still limited
to whatever the kernel itself is configured at, so to really increase
it, one has to modify the kernel as well. Increasing this number allows
for more connections to remain alive while they wait for it, while a
small number will begin rejecting new connections outright as soon as
the limit is reached.